requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
Markdown==3.4.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
schedule==1.1.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14